+++ packages/logic-utils/api.d.ts
@@ -9,6 +9,7 @@ import {
   UI5Association,
   UI5SemanticModel,
   UI5Type,
+  UI5DeprecatedInfo,
 } from "@ui5-language-assistant/semantic-model-types";
 
 /**
@@ -16,6 +17,12 @@ import {
  * and UI5 semantics.
  */
 export function xmlToFQN(astElement: XMLElement): string;
+
+/**
+ * Resolve and XML closing tag's fully qualified name using available `xmlns`
+ * and UI5 semantics.
+ */
+export function xmlClosingTagToFQN(astElement: XMLElement): string;
 
 /**
  * Returns the fully qualified name of a UI5 Semantic Model Node.
@@ -95,6 +102,16 @@ export function getUI5ClassByXMLElement(
 ): UI5Class | undefined;
 
 /**
+ * Return the UI5 Class for an XML Element closing tag
+ * @param element
+ * @param model
+ */
+export function getUI5ClassByXMLElementClosingTag(
+  element: XMLElement,
+  model: UI5SemanticModel
+): UI5Class | undefined;
+
+/**
  * Return the UI5 Aggregation for an XML Element
  * @param element
  * @param model
@@ -128,3 +145,97 @@ export function getUI5NodeFromXMLElementNamespace(
   isDefault: boolean;
   isXmlnsDefined: boolean;
 };
+
+/**
+ * Get the deprecated message. The returned string contains jsdoc tags.
+ *
+ * @param opts.title - The opening sentence for the deprecation message. It should not contain a dot. "Deprecated" by default.
+ */
+export function getDeprecationMessage(opts: {
+  title?: string;
+  since: string | undefined;
+  text: string | undefined;
+}): string;
+
+/**
+ * Get a snippet (first line) of the deprecated documentation without jsdoc tags.
+ *
+ * @param opts.title - The opening sentence for the deprecation message. It should not contain a dot. "Deprecated" by default.
+ */
+export function getDeprecationPlainTextSnippet(opts: {
+  title?: string;
+  deprecatedInfo: UI5DeprecatedInfo;
+  model: UI5SemanticModel;
+}): string;
+
+/**
+ * Convert jsdoc description to markdown format string
+ * @param jsdocDescription
+ * @param model
+ */
+export function convertJSDocToMarkdown(
+  jsdocDescription: string,
+  model: UI5SemanticModel
+): string;
+
+/**
+ * Get a link according to the link text.
+ * Supported links:
+ * - http/https links - returned as-is
+ * - Other strings are considered to be UI5 FQNs. A link to the relevant SDK page is returned.
+ *
+ * @param model
+ * @param link
+ */
+export function getLink(model: UI5SemanticModel, link: string): string;
+
+/**
+ * Split possibly qualified XML Tag or XML Attribute name to prefix and local name.
+ * If there is no prefix in the qualified name, the returned prefix will be undefined.
+ * @param qName
+ */
+export function splitQNameByNamespace(
+  qName: string
+): { prefix: string | undefined; localName: string };
+
+/**
+ * Return the xml namespace defined for the xml element prefix (ns), or undefined if not found
+ * @param xmlElement
+ */
+export function resolveXMLNS(xmlElement: XMLElement): string | undefined;
+
+/**
+ * Return the xml namespace defined for this prefix, or undefined if not found.
+ * The defined namespaces are taken from the xml element.
+ * @param prefix
+ * @param xmlElement
+ */
+export function resolveXMLNSFromPrefix(
+  prefix: string | undefined,
+  xmlElement: XMLElement
+): string | undefined;
+
+/**
+ * Check if the xml element namespace prefixes (ns) reference the same namespace
+ * @param xmlElement1
+ * @param xmlElement2
+ */
+export function isSameXMLNS(
+  xmlElement1: XMLElement,
+  xmlElement2: XMLElement
+): boolean;
+
+/**
+ * Check if the xml namespace prefixes reference the same namespace.
+ * The defined namespaces are taken from the respective xml elements.
+ * @param prefix1
+ * @param xmlElement1
+ * @param prefix2
+ * @param xmlElement2
+ */
+export function isSameXMLNSFromPrefix(
+  prefix1: string | undefined,
+  xmlElement1: XMLElement,
+  prefix2: string | undefined,
+  xmlElement2: XMLElement
+): boolean;