+++ packages/xml-views-tooltip/test/tooltip-spec.ts
@@ -1,7 +1,7 @@
 import { expect } from "chai";
 import { TextDocument, Position } from "vscode-languageserver";
 import { parse, DocumentCstNode } from "@xml-tools/parser";
-import { getAstNodeInPosition } from "@xml-tools/ast-position";
+import { astPositionAtOffset } from "@xml-tools/ast-position";
 import { buildAst } from "@xml-tools/ast";
 import {
   UI5SemanticModel,
@@ -24,9 +24,9 @@ describe("the UI5 language assistant Hover Tooltip Service", () => {
       const xmlSnippet = `<mvc:View
                             xmlns:mvc="sap.ui.core.mvc"
                             xmlns="sap.m"> 
-                            <content>
+                            <mvc:content>
                               <List showSeparator⇶s="All"></List>
-                            </content>
+                            </mvc:content>
                           </mvc:View>`;
       const ui5Node = getUI5Node(xmlSnippet, ui5SemanticModel);
       expectExists(ui5Node, "UI5Node");
@@ -38,24 +38,24 @@ describe("the UI5 language assistant Hover Tooltip Service", () => {
       const xmlSnippet = `<mvc:View
                             xmlns:mvc="sap.ui.core.mvc"
                             xmlns="sap.m"> 
-                            <content>
+                            <mvc:content>
                               <List showSeparator⇶s1="All"></List>
-                            </content>
+                            </mvc:content>
                           </mvc:View>`;
       const ui5Node = getUI5Node(xmlSnippet, ui5SemanticModel);
-      expect(ui5Node).to.not.exist;
+      expect(ui5Node).to.be.undefined;
     });
 
     it("will get hover content UI5 property in unknown tag", () => {
       const xmlSnippet = `<mvc:View
                             xmlns:mvc="sap.ui.core.mvc"
                             xmlns="sap.m"> 
-                            <content>
+                            <mvc:content>
                               <List1 showSeparator⇶s="All"></List1>
-                            </content>
+                            </mvc:content>
                           </mvc:View>`;
       const ui5Node = getUI5Node(xmlSnippet, ui5SemanticModel);
-      expect(ui5Node).to.not.exist;
+      expect(ui5Node).to.be.undefined;
     });
   });
 
@@ -64,9 +64,9 @@ describe("the UI5 language assistant Hover Tooltip Service", () => {
       const xmlSnippet = `<mvc:View
                             xmlns:mvc="sap.ui.core.mvc"
                             xmlns="sap.m"> 
-                            <content>
+                            <mvc:content>
                               <List showSeparators="Al⇶l"></List>
-                            </content>
+                            </mvc:content>
                           </mvc:View>`;
       const ui5Node = getUI5Node(xmlSnippet, ui5SemanticModel);
       expectExists(ui5Node, "UI5Node");
@@ -78,33 +78,33 @@ describe("the UI5 language assistant Hover Tooltip Service", () => {
       const xmlSnippet = `<mvc:View
                             xmlns:mvc="sap.ui.core.mvc"
                             xmlns="sap.m"> 
-                            <content>
+                            <mvc:content>
                               <List busy="fal⇶se"></List>
-                            </content>
+                            </mvc:content>
                           </mvc:View>`;
       const ui5Node = getUI5Node(xmlSnippet, ui5SemanticModel);
-      expect(ui5Node).to.not.exist;
+      expect(ui5Node).to.be.undefined;
     });
 
     it("will get hover content UI5 property - incorrect enum", () => {
       const xmlSnippet = `<mvc:View
                             xmlns:mvc="sap.ui.core.mvc"
                             xmlns="sap.m"> 
-                            <content>
+                            <mvc:content>
                               <List1 showSeparators="Al⇶l1"></List1>
-                            </content>
+                            </mvc:content>
                           </mvc:View>`;
       const ui5Node = getUI5Node(xmlSnippet, ui5SemanticModel);
-      expect(ui5Node).to.not.exist;
+      expect(ui5Node).to.be.undefined;
     });
 
     it("will get hover content UI5 namespace", () => {
       const xmlSnippet = `<mvc:View
                             xmlns:mvc="sap.ui.core.mvc"
                             xmlns="sa⇶p.m"> 
-                            <content>
+                            <mvc:content>
                               <List showSeparators="All"></List>
-                            </content>
+                            </mvc:content>
                           </mvc:View>`;
       const ui5Node = getUI5Node(xmlSnippet, ui5SemanticModel);
       expectExists(ui5Node, "UI5Node");
@@ -118,9 +118,9 @@ describe("the UI5 language assistant Hover Tooltip Service", () => {
       const xmlSnippet = `<mvc:View
                             xmlns:mvc="sap.ui.core.mvc"
                             xmlns="sap.m"> 
-                            <content>
+                            <mvc:content>
                               <L⇶ist showSeparators="All"></List>
-                            </content>
+                            </mvc:content>
                           </mvc:View>`;
       const ui5Node = getUI5Node(xmlSnippet, ui5SemanticModel);
       expectExists(ui5Node, "UI5Node");
@@ -128,13 +128,21 @@ describe("the UI5 language assistant Hover Tooltip Service", () => {
       expect(ui5Node.kind).equal("UI5Class");
     });
 
+    it("will not get hover content UI5 class with a dot", () => {
+      const xmlSnippet = `<core:mvc.Vi⇶ew
+                            xmlns:core="sap.ui.core"> 
+                          </core:mvc.View>`;
+      const ui5Node = getUI5Node(xmlSnippet, ui5SemanticModel);
+      expect(ui5Node).to.be.undefined;
+    });
+
     it("will get hover content UI5 Aggregation", () => {
       const xmlSnippet = `<mvc:View
                             xmlns:mvc="sap.ui.core.mvc"
                             xmlns="sap.m"> 
-                            <conten⇶t>
+                            <mvc:conten⇶t>
                               <List showSeparators="All"></List>
-                            </content>
+                            </mvc:content>
                           </mvc:View>`;
       const ui5Node = getUI5Node(xmlSnippet, ui5SemanticModel);
       expectExists(ui5Node, "UI5Node");
@@ -142,27 +150,79 @@ describe("the UI5 language assistant Hover Tooltip Service", () => {
       expect(ui5Node.kind).equal("UI5Aggregation");
     });
 
-    it("will get hover content of unknown tag with unknown parent tag", () => {
+    it("will get hover content UI5 Aggregation with a different namespace prefix that references the same namespace", () => {
       const xmlSnippet = `<mvc:View
                             xmlns:mvc="sap.ui.core.mvc"
+                            xmlns:mvc2="sap.ui.core.mvc"
                             xmlns="sap.m"> 
-                            <content1>
-                              <List1></Lis⇶t1>
-                            </content1>
+                            <mvc2:conten⇶t>
+                              <List showSeparators="All"></List>
+                            </mvc2:content>
                           </mvc:View>`;
       const ui5Node = getUI5Node(xmlSnippet, ui5SemanticModel);
-      expect(ui5Node).to.not.exist;
+      expectExists(ui5Node, "UI5Node");
+      expect(ui5Node.name).to.equal("content");
+      expect(ui5Node.kind).equal("UI5Aggregation");
+    });
+
+    it("will get hover content UI5 Aggregation in the default namespace", () => {
+      const xmlSnippet = `<View
+                            xmlns="sap.ui.core.mvc"
+                            xmlns:m="sap.m"> 
+                            <conten⇶t>
+                              <m:List showSeparators="All"></m:List>
+                            </content>
+                          </View>`;
+      const ui5Node = getUI5Node(xmlSnippet, ui5SemanticModel);
+      expectExists(ui5Node, "UI5Node");
+      expect(ui5Node.name).to.equal("content");
+      expect(ui5Node.kind).equal("UI5Aggregation");
+    });
+
+    it("will not get hover content for UI5 Aggregation in the wrong namespace", () => {
+      const xmlSnippet = `<mvc:View
+                            xmlns:mvc="sap.ui.core.mvc"
+                            xmlns:m="sap.m"> 
+                            <m:conten⇶t>
+                              <m:List showSeparators="All"></m:List>
+                            </m:content>
+                          </mvc:View>`;
+      const ui5Node = getUI5Node(xmlSnippet, ui5SemanticModel);
+      expect(ui5Node).to.be.undefined;
+    });
+
+    it("will not get hover content UI5 Aggregation when only the aggregation doesn't have a namespace", () => {
+      const xmlSnippet = `<mvc:View
+                            xmlns:mvc="sap.ui.core.mvc"
+                            xmlns:m="sap.m"> 
+                            <conten⇶t>
+                            </content>
+                          </mvc:View>`;
+      const ui5Node = getUI5Node(xmlSnippet, ui5SemanticModel);
+      expect(ui5Node).to.be.undefined;
     });
   });
 
   context("hover on element close tag name", () => {
+    it("will get hover content of unknown tag with unknown parent tag", () => {
+      const xmlSnippet = `<mvc:View
+                            xmlns:mvc="sap.ui.core.mvc"
+                            xmlns="sap.m"> 
+                            <mvc:content1>
+                              <List1></Lis⇶t1>
+                            </mvc:content1>
+                          </mvc:View>`;
+      const ui5Node = getUI5Node(xmlSnippet, ui5SemanticModel);
+      expect(ui5Node).to.be.undefined;
+    });
+
     it("will get hover content UI5 class", () => {
       const xmlSnippet = `<mvc:View
                             xmlns:mvc="sap.ui.core.mvc"
                             xmlns="sap.m"> 
-                            <content>
+                            <mvc:content>
                               <List showSeparators="All"></L⇶ist>
-                            </content>
+                            </mvc:content>
                           </mvc:View>`;
       const ui5Node = getUI5Node(xmlSnippet, ui5SemanticModel);
       expectExists(ui5Node, "UI5Node");
@@ -174,10 +234,39 @@ describe("the UI5 language assistant Hover Tooltip Service", () => {
       const xmlSnippet = `<mvc:View
                             xmlns:mvc="sap.ui.core.mvc"
                             xmlns="sap.m"> 
-                            <content>
+                            <mvc:content>
                               <List showSeparators="All"></List>
-                            </conte⇶nt>
+                            </mvc:conte⇶nt>
                           </mvc:View>`;
+      const ui5Node = getUI5Node(xmlSnippet, ui5SemanticModel);
+      expectExists(ui5Node, "UI5Node");
+      expect(ui5Node.name).to.equal("content");
+      expect(ui5Node.kind).equal("UI5Aggregation");
+    });
+
+    it("will get hover content UI5 Aggregation with a different namespace prefix that references the same namespace", () => {
+      const xmlSnippet = `<mvc:View
+                            xmlns:mvc="sap.ui.core.mvc"
+                            xmlns:mvc2="sap.ui.core.mvc"
+                            xmlns="sap.m"> 
+                            <mvc_TYPO:content>
+                              <List showSeparators="All"></List>
+                            </mvc2:conte⇶nt>
+                          </mvc:View>`;
+      const ui5Node = getUI5Node(xmlSnippet, ui5SemanticModel);
+      expectExists(ui5Node, "UI5Node");
+      expect(ui5Node.name).to.equal("content");
+      expect(ui5Node.kind).equal("UI5Aggregation");
+    });
+
+    it("will get hover content UI5 Aggregation in the default namespace", () => {
+      const xmlSnippet = `<View
+                            xmlns="sap.ui.core.mvc"
+                            xmlnsLm="sap.m"> 
+                            <content>
+                              <m:List showSeparators="All"></m:List>
+                            </conte⇶nt>
+                          </View>`;
       const ui5Node = getUI5Node(xmlSnippet, ui5SemanticModel);
       expectExists(ui5Node, "UI5Node");
       expect(ui5Node.name).to.equal("content");
@@ -195,20 +284,29 @@ describe("the UI5 language assistant Hover Tooltip Service", () => {
       expect(ui5Node.kind).equal("UI5Class");
     });
 
+    it("will not get hover content UI5 class with a dot", () => {
+      const xmlSnippet = `<mvc:View
+                            xmlns:mvc="sap.ui.core.mvc"
+                            xmlns:core="sap.ui.core"> 
+                          </core:mvc.Vi⇶ew>`;
+      const ui5Node = getUI5Node(xmlSnippet, ui5SemanticModel);
+      expect(ui5Node).to.be.undefined;
+    });
+
     it("will not get hover content when only the close tag has the wrong namespace", () => {
       const xmlSnippet = `<mvc:View
                             xmlns:mvc="sap.ui.core.mvc"
                             xmlns="sap.m"> 
                           </m:Vie⇶w>`;
       const ui5Node = getUI5Node(xmlSnippet, ui5SemanticModel);
-      expect(ui5Node).to.not.exist;
+      expect(ui5Node).to.be.undefined;
     });
 
     it("will get hover content UI5 class when open and close tag are different", () => {
       const xmlSnippet = `<mvc:View
                             xmlns:mvc="sap.ui.core.mvc"
                             xmlns="sap.m"> 
-                            <content></L⇶ist>
+                            <mvc:content></L⇶ist>
                           </mvc:View>`;
       const ui5Node = getUI5Node(xmlSnippet, ui5SemanticModel);
       expectExists(ui5Node, "UI5Node");
@@ -221,12 +319,12 @@ describe("the UI5 language assistant Hover Tooltip Service", () => {
     const xmlSnippet = `<mvc:View
                           xmlns:mvc="sap.ui.core.mvc"
                           xmlns="sap.m"> 
-                          <content>
+                          <mvc:content>
                             <List ⇶ showSeparators="All"></List>
-                          </content>
+                          </mvc:content>
                         </mvc:View>`;
     const ui5Node = getUI5Node(xmlSnippet, ui5SemanticModel);
-    expect(ui5Node).to.not.exist;
+    expect(ui5Node).to.be.undefined;
   });
 });
 
@@ -238,7 +336,7 @@ function getUI5Node(
   const { cst, tokenVector } = parse(document.getText());
   const ast = buildAst(cst as DocumentCstNode, tokenVector);
   const offset = document.offsetAt(position);
-  const astPosition = getAstNodeInPosition(ast, offset);
+  const astPosition = astPositionAtOffset(ast, offset);
   if (astPosition !== undefined) {
     const ui5Node = findUI5HoverNodeAtOffset(astPosition, ui5SemanticModel);
     return ui5Node;