+++ packages/xml-views-validation/src/utils/special-namespaces.ts
@@ -4,3 +4,7 @@ export const TEMPLATING_NS =
 /* istanbul ignore next - the constants are used but coverage is not reported for some reason */
 export const CUSTOM_DATA_NS =
   "http://schemas.sap.com/sapui5/extension/sap.ui.core.CustomData/1";
+/* istanbul ignore next - the constants are used but coverage is not reported for some reason */
+export const XHTML_NS = "http://www.w3.org/1999/xhtml";
+/* istanbul ignore next - the constants are used but coverage is not reported for some reason */
+export const SVG_NS = "http://www.w3.org/2000/svg";