+++ packages/logic-utils/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.2.1](https://github.com/sap/ui5-language-assistant/compare/@ui5-language-assistant/logic-utils@1.2.0...@ui5-language-assistant/logic-utils@1.2.1) (2020-06-03)
+
+### Bug Fixes
+
+- don't allow to use classes from sub-namespace in tag ([#155](https://github.com/sap/ui5-language-assistant/issues/155)) ([cb96c80](https://github.com/sap/ui5-language-assistant/commit/cb96c80890f74f8d4ce90648ea6b1f2d915bfccc))
+- support namespace in aggregation name ([#150](https://github.com/sap/ui5-language-assistant/issues/150)) ([cff718b](https://github.com/sap/ui5-language-assistant/commit/cff718b4a2cfddc01cc5e44bd42eca68a8831832))
+- use description first line without jsdoc tags in deprecation warning ([#141](https://github.com/sap/ui5-language-assistant/issues/141)) ([9cf89eb](https://github.com/sap/ui5-language-assistant/commit/9cf89ebda9dbf80c00b499e66cb44fabeb4d3553))
+
 # [1.2.0](https://github.com/sap/ui5-language-assistant/compare/@ui5-language-assistant/logic-utils@1.1.3...@ui5-language-assistant/logic-utils@1.2.0) (2020-05-20)
 
 ### Features